:art:Style: 로그인 관련 함수 signin.js로 이동 #10
index.js에 있던 setLocalStorage 함수와 소셜로그인 시 url에서 access 토큰
얻는 코드 signin.js로 옮겼습니다.
